fix: move department data to separate file for reusability
