+++ app/(app)/dashboard/page.tsx
@@ -11,7 +11,7 @@ import {
   PiStackPlus,
 } from "react-icons/pi";
 import { projects as allProjectsData } from "@/app/(app)/projects/page";
-import { departments } from "../departments/page";
+import { departments } from "../departments/data";
 import { sampleProducts } from "../products/page";
 
 const departmentsLookup: { [key: string]: string } = {
@@ -49,7 +49,7 @@ type DashboardProject = {
 };
 
 // Map the recent projects to the required structure
-export const projects: DashboardProject[] = recentProjectsData.map((p) => ({
+const projects: DashboardProject[] = recentProjectsData.map((p) => ({
   id: p.id,
   image: p.image,
   projectNumber: p.id, // Use project ID as project number for now
