+++ features/source-files/DialogAddProductFolder.tsx
@@ -24,7 +24,7 @@ import {
 } from "@/components/ui/select";
 
 // Import actual data
-import { departments } from "@/app/(app)/departments/page";
+import { departments } from "@/app/(app)/departments/data";
 import { sampleProducts } from "@/app/(app)/products/page";
 import { projects } from "@/app/(app)/projects/page";
 import {
